pyproject.toml
[project]
name = "beep-me-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi==23.1.0",
    "authlib>=1.6.5",
    "boto3>=1.40.50",
    "channels==4.2.2",
    "channels-redis==4.2.1",
    "daphne>=4.2.1",
    "dj-database-url>=3.0.1",
    "django>=5.2",
    "django-cors-headers==4.7.0",
    "django-redis==5.4.0",
    "django-storages>=1.14.6",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "djoser>=2.3.3",
    "drf-yasg==1.21.10",
    "google-auth>=2.41.1",
    "livekit-api==1.0.3",
    "pillow>=11.3.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.1.1",
    "social-auth-app-django>=5.6.0",
    "whitenoise>=6.11.0",
]
